cart, favorites, reviews

+++ src/product/product.service.ts
@@ -1,28 +1,39 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common'
+import {
+	Injectable,
+	InternalServerErrorException,
+	NotFoundException,
+} from '@nestjs/common'
 import { PrismaService } from 'src/prisma.service'
 import { ProductDto } from './dto/product.dto'
 import { GoogleCloudService } from 'src/upload-files/google-cloud.service'
+import { Products } from 'generated/prisma'
 
 @Injectable()
 export class ProductService {
 	constructor(
-		private prismaService: PrismaService,
-		private googleCloudService: GoogleCloudService
+		private readonly prismaService: PrismaService,
+		private readonly googleCloudService: GoogleCloudService
 	) {}
 
-	async getAll() {
-		return this.prismaService.products.findMany({
+	async getAll(): Promise<Products[]> {
+		return await this.prismaService.products.findMany({
 			include: { discount: true, reviews: true },
 		})
 	}
 
-	async getById(id: string) {
-		return this.prismaService.products.findUnique({ where: { id } })
+	async getById(id: string): Promise<Products> {
+		const product = await this.prismaService.products.findUnique({
+				where: { id },
+		})
+
+		if (!product) throw new NotFoundException(`Product not found`)
+
+		return product
 	}
 
-	async create(data: ProductDto, urls: string[]) {
+	async create(data: ProductDto, urls: string[]): Promise<Products> {
 		try {
-			return this.prismaService.products.create({
+			const newProduct = await this.prismaService.products.create({
 				data: {
 					...data,
 					gasoline: +data.gasoline,
@@ -30,29 +41,39 @@ export class ProductService {
 					icons: urls,
 				},
 			})
+			return newProduct
 		} catch (error) {
 			await this.googleCloudService.deleteFiles(urls)
-			throw new InternalServerErrorException('Invalid to create a product')
+			throw new InternalServerErrorException('Failed to create product')
 		}
 	}
 
-	async deleteAll() {
-		try {
-			const result = await this.prismaService.products.findMany({
-				select: { icons: true },
+	async deleteAll(): Promise<{ message: string }> {
+		const products = await this.prismaService.products.findMany({
+			select: { icons: true },
+		})
+
+		if(products.length === 0) throw new NotFoundException('No products found to delete')
+
+		const allIcons = products.flatMap((item) => item.icons)
+
+		await this.prismaService.products.deleteMany()
+		await this.googleCloudService.deleteFiles(allIcons).catch(() => {
+			throw new InternalServerErrorException('Failed to delete icons')
+		})
+
+		return { message: 'Products were deleted successfully' }
+	}
+
+	async deleteById(id: string): Promise<{ message: string }> {
+			const product = await this.getById(id)
+
+			await this.prismaService.products.delete({ where: { id } })
+
+			await this.googleCloudService.deleteFiles(product.icons).catch(() => {
+				throw new InternalServerErrorException('Failed to delete icon')
 			})
-			const urls = result.flatMap((item) => item.icons)
 
-			await this.prismaService.products.deleteMany()
-			await this.googleCloudService.deleteFiles(urls)
-		} catch (error) {
-			console.log(error)
-			throw new InternalServerErrorException(
-				'Invalid to delete products'
-			)
-		}
-	}
-	async deleteById(id: string) {
-		return this.prismaService.products.delete({ where: { id } })
+			return { message: `Product deleted successfully` }
 	}
 }
